% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_csv_server.R
\name{csvFilePathServer}
\alias{csvFilePathServer}
\title{Module server function}
\usage{
csvFilePathServer(id, stringsAsFactors)
}
\value{
a module server that can read a csv file
}
\description{
Module server function
}
