% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_user_management.R
\name{lookup_id_from_name}
\alias{lookup_id_from_name}
\title{user_id of a valid name string}
\usage{
lookup_id_from_name(name)
}
\arguments{
\item{name}{a string representation of the name you want to look up}
}
\value{
user_id user ID from \code{user_df_from_libreview}
}
\description{
Assuming the name string is already in the user database, returns the user_id
}
