% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_filter_glucose_server}
\alias{mod_filter_glucose_server}
\title{Shiny server to return a valid glucose dataframe}
\usage{
mod_filter_glucose_server(id, user_id = 1234)
}
\value{
dataframe
}
\description{
Shiny server to return a valid glucose dataframe
}
