% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filterServer}
\alias{mod_filterServer}
\title{Shiny module (server) to filter a CGM result based on user input}
\usage{
mod_filterServer(id)
}
\arguments{
\item{id}{server input id}
}
\value{
valid libreview-based dataframe encapsulated as a reactive
}
\description{
Shiny module (server) to filter a CGM result based on user input
}
