% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filterUI}
\alias{mod_filterUI}
\title{Shiny module (UI) to filter CGM result based on user input}
\usage{
mod_filterUI(id)
}
\arguments{
\item{id}{input id for server}
}
\description{
Standalone UI renders HTML for a main panel
}
