% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_food2UI.R
\name{mod_food2Server}
\alias{mod_food2Server}
\title{Shiny server to display two food comparisons for one user.}
\usage{
mod_food2Server(id, glucose_df, title = "Name")
}
\arguments{
\item{id}{shiny module id}

\item{glucose_df}{reactive for a valid glucose dataframe}

\item{title}{a title for the plot}
}
\value{
ggplot object representing a glucose chart
}
\description{
Given a (reactive) libreview dataframe, this Shiny module will
generate a valid ggplot object and display it in an accompanying UI
}
