% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_food2UI.R
\name{mod_food2UI}
\alias{mod_food2UI}
\title{UI for comparing two foods for a single user}
\usage{
mod_food2UI(id)
}
\arguments{
\item{id}{Shiny id}
}
\description{
Plot glucose responses for two foods for a single user.
}
