% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_libreviewUI.R
\name{mod_libreview_plotServer}
\alias{mod_libreview_plotServer}
\title{Make a glucose chart}
\usage{
mod_libreview_plotServer(id, glucose_df, title = "Name")
}
\arguments{
\item{id}{shiny module id}

\item{glucose_df}{reactive for a valid glucose dataframe}

\item{title}{a title for the plot}
}
\value{
ggplot object representing a glucose chart
}
\description{
Given a (reactive) libreview dataframe, this Shiny module will
generate a valid ggplot object and display it in an accompanying UI
}
