% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_libreviewUI.R
\name{mod_libreviewUI}
\alias{mod_libreviewUI}
\title{UI for Libreview plots}
\usage{
mod_libreviewUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\description{
Plot a Libreview object along with its AUC value
}
