% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{normalize_value}
\alias{normalize_value}
\title{return a new df where value are normalized to start from zero.}
\usage{
normalize_value(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
return a new df where value are normalized to start from zero.
}
