% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_plot.R
\name{plot_food_compare}
\alias{plot_food_compare}
\title{plot differences in glucose response to a particular food}
\usage{
plot_food_compare(
  food_times = food_times_df(foodname = "watermelon"),
  foodname = "watermelon"
)
}
\arguments{
\item{food_times}{a dataframe of food times, preferably generated by \code{food_times_df()}}

\item{foodname}{character string matching the food of interest}
}
\value{
ggplot object
}
\description{
Given a dataframe of food times, returns a ggplot object
representing the plot of all different food times along a constant time period
}
