% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_plot.R
\docType{data}
\name{psi_theme}
\alias{psi_theme}
\title{Stylized theme for consistent plot style}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 3.
}
\usage{
psi_theme
}
\description{
Stylized theme for consistent plot style
}
\keyword{datasets}
