% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiCGM-package.R
\docType{package}
\name{psiCGM-package}
\alias{psiCGM}
\alias{psiCGM-package}
\title{psiCGM: Personal Science CGM Analysis Web App}
\description{
Shiny web app for generating reports based on CGM data.
}
\author{
\strong{Maintainer}: Richard Sprague \email{sprague@personalscience.com} (\href{https://orcid.org/0000-0002-3789-5747}{ORCID})

}
\keyword{internal}
