% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Personal Science Web App}
\usage{
run_app(mode = attr(config::get(), "config"))
}
\arguments{
\item{mode}{character string indicating operational mode: Uses config::get() or "run" (default)}
}
\description{
Run the Personal Science CGM Web Application
}
