% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastermonial.R
\name{taster_df}
\alias{taster_df}
\title{Read a Tastermonial Airtable file from CSV}
\usage{
taster_df(
  file = file.path(config::get("tastermonial")$datadir, "TastermonialNotes.csv")
)
}
\arguments{
\item{file}{path to a CSV file}
}
\value{
canonical dataframe of Tastermonial information
}
\description{
Turn a Tastermonial Airtable file into a canonical dataframe.
}
