% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_df_from_libreview}
\alias{user_df_from_libreview}
\title{Table of all registered users}
\format{
a data frame with 6 rows
}
\usage{
data(user_df_from_libreview)
}
\description{
from a CSV file generated at the Libreview Practice Portal
}
\keyword{datasets}
